Move request code from ViewController to OpenWeatherMapClient in Model group

+++ Sunshine/ForecastDetail.swift
@@ -13,13 +13,13 @@ struct ForecastDetail{
     let shortDes: String
     let maxTemp: Int
     let minTemp: Int
-    let humidity: Float
+    let humidity: Int
     let pressure: Float
     let windSpeed: Float
     
     
     //MARK: Initializers
-    init(dateStr: String, weatherCode: Int, shortDes: String, maxTemp: Int, minTemp: Int, humidity: Float, pressure: Float, windSpeed: Float){
+    init(dateStr: String, weatherCode: Int, shortDes: String, maxTemp: Int, minTemp: Int, humidity: Int, pressure: Float, windSpeed: Float){
         self.dateStr = dateStr
         self.weatherCode = weatherCode
         self.shortDes = shortDes
